XStream- Object to Xml
Closing Review comments, test cases added.
