+++ xstream-introduction/src/main/java/com/baeldung/pojo/AddressDetails.java
@@ -1,4 +1,4 @@
-package org.baeldung.pojo;
+package com.baeldung.pojo;
 
 import java.util.List;
