+++ xstream-introduction/src/main/java/com/baeldung/pojo/ContactDetails.java
@@ -1,4 +1,4 @@
-package org.baeldung.pojo;
+package com.baeldung.pojo;
 
 import com.thoughtworks.xstream.annotations.XStreamAlias;
